requirements.txt
# Core project dependencies
ultralytics>=8.3.0
supervision>=0.15.0
torch>=1.10.0  # Used indirectly by YOLO (Ultralytics)
umap-learn>=0.5.3
transformers
sentencepiece
protobuf

# Data manipulation
numpy>=1.21.0
pandas>=1.3.0

# Computer vision
opencv-python>=4.5.3

# Visualization and progress bars
tqdm>=4.62.0
matplotlib>=3.4.3

# Hydra for config management
hydra-core>=1.1.0
omegaconf>=2.1.0

# Others (custom modules in your project rely on these)
scipy>=1.7.0